+++ lib/internal/modules/cjs/loader.js
@@ -218,16 +218,6 @@ function tryExtensions(p, exts, isMain) {
   return false;
 }
 
-function readExtensions() {
-  const exts = Object.keys(Module._extensions);
-  for (var i = 0, j = 0; i < exts.length; ++i) {
-    if (path.extname(exts[i]) === '')
-      exts[j++] = exts[i];
-  }
-  exts.length = j;
-  return exts;
-}
-
 var warned = false;
 Module._findPath = function(request, paths, isMain) {
   if (path.isAbsolute(request)) {
@@ -284,7 +274,7 @@ Module._findPath = function(request, paths, isMain) {
       if (!filename) {
         // try it with each of the extensions
         if (exts === undefined)
-          exts = readExtensions();
+          exts = Object.keys(Module._extensions);
         filename = tryExtensions(basePath, exts, isMain);
       }
     }
@@ -292,7 +282,7 @@ Module._findPath = function(request, paths, isMain) {
     if (!filename && rc === 1) {  // Directory.
       // try it with each of the extensions at "index"
       if (exts === undefined)
-        exts = readExtensions();
+        exts = Object.keys(Module._extensions);
       filename = tryPackage(basePath, exts, isMain);
       if (!filename) {
         filename = tryExtensions(path.resolve(basePath, 'index'), exts, isMain);
